Refactor install strategy package into multiple files
